+++ changelog/7268.fixed.md
@@ -1,0 +1,1 @@
+Fixed missing changelog events for objects created as side effects during node creation, including pool-allocated resources and template-generated child objects.
